MINOR: mux-h2/traces: also suggest invalid header upon parsing error
Historically the parsing error used to apply only to too large headers,
so this is what has been reported in traces. But nowadays we can also
reject invalid characters, and when this happens the trace is a bit
misleading, so let's mention "or invalid".

(cherry picked from commit db97bb42d9b435846809a028dce077eea1621b40)
Signed-off-by: Willy Tarreau <w@1wt.eu>