BUG/MEDIUM: dns: Be sure to unlock DSS when existing dns_session_io_handler()
A regression was introduced when the commit cf537f687 ("BUG/MEDIUM: dns:
Properly handle error when a response consumed") was backported. When the
dns session I/O handler is called, the DSS lock is not acquired in upper
versions. But in the 2.4, DSS may be locked. So we must be sure to never
exit from the I/O handler without unlocking it.

This patch should fix the issue #2277. It is specific to 2.4, thus there is
no upstream ID. No backport needed.