MINOR: httpclient: allow to configure the retries
When using the httpclient, one could be bothered with it returning after
a very long time when failing. By default the httpclient has a retries
of 3 and a timeout connect of 5s, which can results in pause of 20s
upon failure.

This patch allows the user to configure the retries of the httpclient so
it could reduce the time to return an error.

This patch helps fixing part of the issue #2269.

Could be backported in 2.7 if needed.

(cherry picked from commit c52948bd2cad65cdee62744d315686cfa317125f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 4ad693ed2bf5d7b1a04b499a2db0e2e78cbf5f41)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>